+++ Forge/changelog.md
@@ -1,3 +1,6 @@
-1.0.14:
-- Added Quark Parrot Eggs
-- ChocoCraft integration
+1.0.15:
+- Ice and Fire Integration
+- Aether Integration
+- Aether: Redux Integration
+- Deep Aether Integration
+- Implements [#11](https://github.com/Christofmeg/JustEnoughBreeding/issues/11)